fixing bugs from educations.controller and creating educations.router

+++ src/backend/controllers/educations.controller.js
@@ -8,7 +8,7 @@ class EducationsController {
 
     createEducation = async (req, res) => {
         const data = req.body;
-        if(!data || !data.institutionName || !data.title || !data.dateStart || !data.typeEducation || !data.description ) { 
+        if (!data || !data.institutionName || !data.title || !data.dateStart || !data.typeEducation || !data.description) {
             res.json400("Missing Information!(C)");
         };
         const education = await this.eService.createOne(data);
@@ -18,8 +18,9 @@ class EducationsController {
 
     getEducationById = async (req, res) => {
         const { eid } = req.params;
-        const education = this.verifyEducationFun(eid);
-        if(education === null) { return res.json400("Education Not Found!(C)"); };
+        if (eid.length !== 24) { return res.json400("Invalid Education ID!(C)"); };
+        const education = await this.verifyEducationFun(eid);
+        if (education === null) { return res.json400("Education Not Found!(C)"); };
         res.json200(education);
     };
 
@@ -28,23 +29,26 @@ class EducationsController {
     }
 
     getAllEducations = async (req, res) => {
-        const educations = this.eService.getAllEducations();
+        const educations = await this.eService.readAll();
         if (educations.length === 0) { return res.json404("Not Educations found!(C)"); };;
         return res.json200(educations);
     };
 
     updateEducationById = async (req, res) => {
         const { eid } = req.params;
+        if (eid.length !== 24) { return res.json400("Invalid Education ID!(C)"); };
         const data = req.body;
         if (!data) { return res.json400("No data to update!(C)"); };
-        const education = this.verifyEducationFun(eid);
+        const education = await this.verifyEducationFun(eid);
+        //FALTA VERIFICAR SI YA EXISTE LA EDUCACION POR OTROS FILTROS
         if (education === null) { return res.json404("No Education found!(C)"); };
-        const educationUpdated = this.eService.updateById(eid, data);
+        const educationUpdated = await this.eService.updateById(eid, data);
         return res.json200(educationUpdated);
     };
 
     deleteEducationById = async (req, res) => {
         const { eid } = req.params;
+        if (eid.length !== 24) { return res.json400("Invalid Education ID!(C)"); };
         const education = await this.verifyEducationFun(eid);
         if (education === null) { return res.json404("Education Not Found!(C)"); };
         const educationDeleted = await this.eService.destroyById(eid);
@@ -52,9 +56,9 @@ class EducationsController {
     };
 
     verifyEducationFun = async (eid) => {
-        if(!isValidObjectId(eid)) { return res.json400("Invalid Education ID!(C)"); };
-        const verify = await this.eService.getEducationById(eid);
-        if(!verify) { return null }
+        if (!isValidObjectId(eid)) { return res.json400("Invalid Education ID!(C)"); };
+        const verify = await this.eService.readById(eid);
+        if (!verify) { return null }
         else { return verify; };
     };
 };
